+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Small refactoring
 - Return path of downloaded file
 - Move exceptions to separate directory
+- Make destination nullable
 
 ## [0.1.0] - 2022-07-10
 ### Added